ai_server/requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
Flask==2.3.2
requests==2.32.2
numpy==1.26.4
opencv-python==4.10.0.84
pandas==1.5.3
pyjwt==2.6.0
openai-whisper>=20230918
librosa==0.10.2
soundfile==0.12.1
sounddevice==0.5.0
scipy==1.14.1
# TTS>=0.22.0
# TTS 기능 D-ID API에서 제공하므로 주석처리
flask-cors==4.0.0
unidecode==1.3.8
pydub>=0.25.1
numba>=0.56.0
psutil>=5.9.0

# AI 아바타 API 관련 패키지
python-dotenv>=1.0.0
pillow>=10.0.0